@@ -1,10 +1,11 @@
-<?php namespace Tarsana\Filesystem\Adapters;
+<?php
+
+namespace Tarsana\Filesystem\Adapters;
 
 use Tarsana\Filesystem\Interfaces\Adapter;
 
-
-class Local implements Adapter {
-
+class Local implements Adapter
+{
     /**
      * The singleton instance.
      *
@@ -20,7 +21,7 @@ class Local implements Adapter {
     public static function instance()
     {
         if (null === self::$instance) {
-            self::$instance = new Local;
+            self::$instance = new Local();
         }
         return self::$instance;
     }
@@ -28,7 +29,9 @@ class Local implements Adapter {
     /**
      * Priate Constructor.
      */
-    private function __construct() {}
+    private function __construct()
+    {
+    }
 
     /**
      * Tells if the given path is absolute.
@@ -257,8 +260,9 @@ class Local implements Adapter {
     public function createFile($path)
     {
         $created = fopen($path, "w");
-        if (false === $created)
+        if (false === $created) {
             return false;
+        }
 
         fclose($created);
         return true;
@@ -285,5 +289,4 @@ class Local implements Adapter {
     {
         return basename($path);
     }
-
 }
