@@ -1,6 +1,10 @@
-<?php namespace Tarsana\Filesystem\Interfaces\Resource;
+<?php
+
+namespace Tarsana\Filesystem\Interfaces\Resource;
 
 /**
  * Reads/Writes content from/into a stream of data.
  */
-interface Buffer extends Reader, Writer {}
+interface Buffer extends Reader, Writer
+{
+}
