@@ -4,35 +4,32 @@ use Tarsana\Filesystem\Filesystem;
 use Tarsana\Filesystem\Adapters\Local;
 use Tarsana\Filesystem\Directory;
 
-
-class DirectoryTest extends PHPUnit\Framework\TestCase {
-
+class DirectoryTest extends PHPUnit\Framework\TestCase
+{
     protected $dirPath;
 
     protected $dir;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->dirPath = DEMO_DIR.'/folder1';
+        $this->dirPath = DEMO_DIR . '/folder1';
         $this->dir = new Directory($this->dirPath);
     }
 
     public function test_creates_directory_if_missing()
     {
-        $this->assertFalse(is_dir(DEMO_DIR.'/temp'));
+        $this->assertFalse(is_dir(DEMO_DIR . '/temp'));
 
-        $dir = new Directory(DEMO_DIR.'/temp');
-        $this->assertTrue(is_dir(DEMO_DIR.'/temp'));
+        $dir = new Directory(DEMO_DIR . '/temp');
+        $this->assertTrue(is_dir(DEMO_DIR . '/temp'));
 
-        rmdir(DEMO_DIR.'/temp');
+        rmdir(DEMO_DIR . '/temp');
     }
 
-    /**
-     * @expectedException Tarsana\Filesystem\Exceptions\FilesystemException
-     */
     public function test_throws_exception_when_cannot_create_the_directory()
     {
-        $dir = new Directory(DEMO_DIR.'/files.txt');
+        $this->expectException(\Tarsana\Filesystem\Exceptions\FilesystemException::class);
+        $dir = new Directory(DEMO_DIR . '/files.txt');
     }
 
     public function test_gets_filesystem_instance()
@@ -44,10 +41,10 @@ class DirectoryTest extends PHPUnit\Framework\TestCase {
     {
         $this->assertTrue($this->dir->exists());
 
-        $otherDir = new Directory(DEMO_DIR.'/temp');
+        $otherDir = new Directory(DEMO_DIR . '/temp');
         $this->assertTrue($otherDir->exists());
 
-        rmdir(DEMO_DIR.'/temp');
+        rmdir(DEMO_DIR . '/temp');
         $this->assertFalse($otherDir->exists());
     }
 
@@ -55,9 +52,9 @@ class DirectoryTest extends PHPUnit\Framework\TestCase {
     {
         $this->assertEquals($this->dirPath, $this->dir->path());
 
-        $this->dir->path(DEMO_DIR.'/folder2/folder1', true);
-        $this->assertEquals(DEMO_DIR.'/folder2/folder1', $this->dir->path());
-        $this->assertTrue(is_dir(DEMO_DIR.'/folder2/folder1'));
+        $this->dir->path(DEMO_DIR . '/folder2/folder1', true);
+        $this->assertEquals(DEMO_DIR . '/folder2/folder1', $this->dir->path());
+        $this->assertTrue(is_dir(DEMO_DIR . '/folder2/folder1'));
 
         $this->dir->path($this->dirPath);
         $this->assertEquals($this->dirPath, $this->dir->path());
@@ -70,8 +67,8 @@ class DirectoryTest extends PHPUnit\Framework\TestCase {
 
         $this->dir->name('new-folder');
         $this->assertEquals('new-folder', $this->dir->name());
-        $this->assertEquals(DEMO_DIR.'/new-folder', $this->dir->path());
-        $this->assertTrue(is_dir(DEMO_DIR.'/new-folder'));
+        $this->assertEquals(DEMO_DIR . '/new-folder', $this->dir->path());
+        $this->assertTrue(is_dir(DEMO_DIR . '/new-folder'));
 
         $this->dir->name('folder1');
         $this->assertEquals('folder1', $this->dir->name());
@@ -91,11 +88,11 @@ class DirectoryTest extends PHPUnit\Framework\TestCase {
 
     public function test_copy_as()
     {
-        $copy = $this->dir->copyAs(DEMO_DIR.'/copies/folder1');
+        $copy = $this->dir->copyAs(DEMO_DIR . '/copies/folder1');
 
-        $this->assertTrue(is_dir(DEMO_DIR.'/copies/folder1'));
+        $this->assertTrue(is_dir(DEMO_DIR . '/copies/folder1'));
         $this->assertTrue($copy instanceof Directory);
-        $this->assertEquals(DEMO_DIR.'/copies/folder1', $copy->path());
+        $this->assertEquals(DEMO_DIR . '/copies/folder1', $copy->path());
 
         $this->assertEquals(1, $copy->fs()->dirs()->count());
         $this->assertEquals(2, $copy->fs()->files()->count());
@@ -105,19 +102,19 @@ class DirectoryTest extends PHPUnit\Framework\TestCase {
 
     public function test_remove()
     {
-        $dir = new Directory(DEMO_DIR.'/temp');
+        $dir = new Directory(DEMO_DIR . '/temp');
         $this->assertTrue(
-            file_exists(DEMO_DIR.'/temp')
-            && is_dir(DEMO_DIR.'/temp')
+            file_exists(DEMO_DIR . '/temp')
+            && is_dir(DEMO_DIR . '/temp')
         );
 
         $dir->remove();
-        $this->assertFalse(file_exists(DEMO_DIR.'/temp'));
+        $this->assertFalse(file_exists(DEMO_DIR . '/temp'));
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
-        remove(DEMO_DIR.'/temp');
-        remove(DEMO_DIR.'/copies');
+        remove(DEMO_DIR . '/temp');
+        remove(DEMO_DIR . '/copies');
     }
 }
