@@ -3,30 +3,28 @@
 use Tarsana\Filesystem\Adapters\Local;
 use Tarsana\Filesystem\Resource\Writer;
 
-class WriterTest extends PHPUnit\Framework\TestCase {
-
+class WriterTest extends PHPUnit\Framework\TestCase
+{
     protected $writer;
 
     protected $path;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->path = DEMO_DIR.'/temp.txt';
+        $this->path = DEMO_DIR . '/temp.txt';
         file_put_contents($this->path, "");
         $this->writer = new Writer($this->path);
     }
 
-    /**
-     * @expectedException Tarsana\Filesystem\Exceptions\ResourceException
-     */
     public function test_fails_if_not_writable()
     {
+        $this->expectException(\Tarsana\Filesystem\Exceptions\ResourceException::class);
         $writer = new Writer(fopen('php://memory', 'r'));
     }
 
     public function test_constructor()
     {
-        $out = new Writer;
+        $out = new Writer();
         $this->assertTrue($out instanceof Writer);
     }
 
@@ -48,9 +46,8 @@ class WriterTest extends PHPUnit\Framework\TestCase {
         $this->assertEquals("Hello World" . PHP_EOL, file_get_contents($this->path));
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
-        remove(DEMO_DIR.'/temp.txt');
+        remove(DEMO_DIR . '/temp.txt');
     }
-
 }
